+++ examples/configuration_space_obstacle2d.rs
@@ -9,8 +9,8 @@ fn main() {
     let cone       = Cone::new(0.75f32, 0.75);
     let reflection = Reflection::new(&cone); // Take the reflection of the cone.
 
-    let delta_cylinder = na::one::<Isometry2<f32>>(); // identity matrix.
-    let delta_cone     = na::one::<Isometry2<f32>>(); // identity matrix.
+    let delta_cylinder = na::one::<Isometry2<f32>>(); // Identity matrix.
+    let delta_cone     = na::one::<Isometry2<f32>>(); // Identity matrix.
 
     // Build the Configuration Space Obstacle.
     let _ = MinkowskiSum::new(&delta_cylinder, &cylinder, &delta_cone, &reflection);