+++ ncollide_pipeline/narrow_phase/default_narrow_phase.rs
@@ -2,17 +2,17 @@ use utils::data::hash_map::HashMap;
 use utils::data::pair::{Pair, PairTWHash};
 use utils::data::uid_remap::{UidRemap, FastKey};
 use geometry::query::Proximity;
-use narrow_phase::{CollisionDispatcher, CollisionAlgorithm, ContactSignal,   CollisionDetector,
+use narrow_phase::{ContactDispatcher, ContactAlgorithm, ContactSignal,   ContactGenerator,
                    ProximityDispatcher, ProximityAlgorithm, ProximitySignal, ProximityDetector,
                    NarrowPhase, ContactPairs, ProximityPairs};
-use world::{CollisionObject, CollisionQueryType};
+use world::{CollisionObject, GeometricQueryType};
 use math::Point;
 
 // FIXME: move this to the `narrow_phase` module.
 /// Collision detector dispatcher for collision objects.
 pub struct DefaultNarrowPhase<P, M> {
-    collision_dispatcher: Box<CollisionDispatcher<P, M> + 'static>,
-    collision_detectors:  HashMap<Pair, CollisionAlgorithm<P, M>, PairTWHash>,
+    contact_dispatcher: Box<ContactDispatcher<P, M> + 'static>,
+    contact_generators:  HashMap<Pair, ContactAlgorithm<P, M>, PairTWHash>,
 
     proximity_dispatcher: Box<ProximityDispatcher<P, M> + 'static>,
     proximity_detectors:  HashMap<Pair, ProximityAlgorithm<P, M>, PairTWHash>,
@@ -20,12 +20,12 @@ pub struct DefaultNarrowPhase<P, M> {
 
 impl<P: Point, M: 'static> DefaultNarrowPhase<P, M> {
     /// Creates a new `DefaultNarrowPhase`.
-    pub fn new(collision_dispatcher: Box<CollisionDispatcher<P, M> + 'static>,
+    pub fn new(contact_dispatcher: Box<ContactDispatcher<P, M> + 'static>,
                proximity_dispatcher: Box<ProximityDispatcher<P, M> + 'static>)
                -> DefaultNarrowPhase<P, M> {
         DefaultNarrowPhase {
-            collision_dispatcher: collision_dispatcher,
-            collision_detectors:  HashMap::new(PairTWHash::new()),
+            contact_dispatcher: contact_dispatcher,
+            contact_generators:  HashMap::new(PairTWHash::new()),
 
             proximity_dispatcher: proximity_dispatcher,
             proximity_detectors:  HashMap::new(PairTWHash::new())
@@ -39,25 +39,25 @@ impl<P: Point, M: 'static, T> NarrowPhase<P, M, T> for DefaultNarrowPhase<P, M> 
               contact_signal:   &mut ContactSignal<T>,
               proximity_signal: &mut ProximitySignal<T>,
               timestamp:        usize) {
-        for e in self.collision_detectors.elements_mut().iter_mut() {
+        for e in self.contact_generators.elements_mut().iter_mut() {
             let co1 = &objects[e.key.first];
             let co2 = &objects[e.key.second];
 
             if co1.timestamp == timestamp || co2.timestamp == timestamp {
-                let had_colls = e.value.num_colls() != 0;
+                let had_contacts = e.value.num_contacts() != 0;
 
-                e.value.update(&*self.collision_dispatcher,
+                e.value.update(&*self.contact_dispatcher,
                                &co1.position, co1.shape.as_ref(),
                                &co2.position, co2.shape.as_ref(),
                                co1.query_type.query_limit() + co2.query_type.query_limit());
 
-                if e.value.num_colls() == 0 {
-                    if had_colls {
+                if e.value.num_contacts() == 0 {
+                    if had_contacts {
                         contact_signal.trigger_contact_signal(&co1.data, &co2.data, false);
                     }
                 }
                 else {
-                    if !had_colls {
+                    if !had_contacts {
                         contact_signal.trigger_contact_signal(&co1.data, &co2.data, true)
                     }
                 }
@@ -97,20 +97,20 @@ impl<P: Point, M: 'static, T> NarrowPhase<P, M, T> for DefaultNarrowPhase<P, M> 
         let co2 = &objects[*fk2];
 
         match (co1.query_type, co2.query_type) {
-            (CollisionQueryType::Contacts(_), CollisionQueryType::Contacts(_)) => {
+            (GeometricQueryType::Contacts(_), GeometricQueryType::Contacts(_)) => {
                 if started {
-                    let cd = self.collision_dispatcher.get_collision_algorithm(co1.shape.as_ref(), co2.shape.as_ref());
+                    let cd = self.contact_dispatcher.get_contact_algorithm(co1.shape.as_ref(), co2.shape.as_ref());
 
                     if let Some(cd) = cd {
-                        let _ = self.collision_detectors.insert(key, cd);
+                        let _ = self.contact_generators.insert(key, cd);
                     }
                 }
                 else {
                     // Proximity stopped.
-                    match self.collision_detectors.get_and_remove(&key) {
+                    match self.contact_generators.get_and_remove(&key) {
                         Some(detector) => {
                             // Trigger the collision lost signal if there was a contact.
-                            if detector.value.num_colls() != 0 {
+                            if detector.value.num_contacts() != 0 {
                                 contact_signal.trigger_contact_signal(&co1.data, &co2.data, false);
                             }
                         },
@@ -118,7 +118,7 @@ impl<P: Point, M: 'static, T> NarrowPhase<P, M, T> for DefaultNarrowPhase<P, M> 
                     }
                 }
             },
-            (_, CollisionQueryType::Proximity(_)) | (CollisionQueryType::Proximity(_), _) => {
+            (_, GeometricQueryType::Proximity(_)) | (GeometricQueryType::Proximity(_), _) => {
                 if started {
                     let cd = self.proximity_dispatcher.get_proximity_algorithm(co1.shape.as_ref(), co2.shape.as_ref());
 
@@ -147,7 +147,7 @@ impl<P: Point, M: 'static, T> NarrowPhase<P, M, T> for DefaultNarrowPhase<P, M> 
 
     fn contact_pairs<'a>(&'a self, objects: &'a UidRemap<CollisionObject<P, M, T>>)
                          -> ContactPairs<'a, P, M, T> {
-        ContactPairs::new(objects, self.collision_detectors.elements().iter())
+        ContactPairs::new(objects, self.contact_generators.elements().iter())
     }
 
     fn proximity_pairs<'a>(&'a self, objects: &'a UidRemap<CollisionObject<P, M, T>>)